+++ tests/test_factory.py
@@ -1,5 +1,6 @@
 """Tests for the application factory."""
 
+import json
 from unittest.mock import Mock, patch
 
 import pytest
@@ -30,7 +31,37 @@ def test_factory_decorator_with_args():
     mock_db_interface.select_interface.assert_called_once()
 
 
-def test_invalid_application_config():
-    with pytest.raises(TypeError):
-        # This will fail because the `TestingConfig` is not instance-based
-        DryFlask.set_default_config_type(_TestingConfig)
+class TestDryFlask:
+    """Tests for the ``DryFlask`` object."""
+
+    def test_initialize(self):
+        app = DryFlask(__name__)
+        assert app.app_name == __name__
+        assert "dry_foundation" in app.blueprints
+
+    def test_initialize_with_name(self):
+        app_name = "Test Application"
+        app = DryFlask(__name__, app_name)
+        assert app.app_name == app_name
+
+    def test_invalid_application_config(self):
+        with pytest.raises(TypeError):
+            # This will fail because the `TestingConfig` is not instance-based
+            DryFlask.set_default_config_type(_TestingConfig)
+
+    def test_serve_js(self, app, client, client_context):
+        js_url = app.url_for("dry_foundation.static", filename="js/modules/requests.js")
+        response = client.get(js_url)
+        assert response.status_code == 200
+        assert "javascript" in response.content_type
+
+    def test_import_map(self, app, client_context):
+        map_imports = app.jinja_env.globals.get("map_imports")
+        # Confirm the import mapper exists and returns validl JSON
+        assert map_imports
+        import_map = json.loads(map_imports())
+        # Confirm import map has the proper URL
+        expected_url = app.url_for(
+            "dry_foundation.static", filename="js/modules/requests.js"
+        )
+        assert import_map["imports"]["dry-foundation/requests"] == expected_url